+++ src/App.css
@@ -20,3 +20,55 @@
     opacity: 1;
   }
 }
+
+/*voor mobiel */
+html, body {
+  margin: 0;
+  padding: 0;
+  width: 100%;
+  overflow-x: hidden;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+}
+
+.bento-grid, .bento-grid-onderste {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+  max-width: 400px;
+  margin: 0 auto;
+}
+
+.box {
+  width: 100%;
+  max-width: 400px;
+  height: 250px;
+  margin: 0px auto;
+  text-align: center;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+@media (max-width: 768px) {
+  .bento-grid, .bento-grid-onderste {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    width: 100%;
+    max-width: 400px;
+    margin: 0 auto;
+  }
+
+  .box {
+    width: 100%;
+    max-width: 400px;
+    height: 250px;
+    margin: 0px auto;
+  }
+}
