+++ src/Components/BentoBox/BentoBox.css
@@ -59,7 +59,7 @@ body {
         width: 100%;
         height: 250px;
         max-width: 400px;
-        margin: 10px auto;
+        margin: 0px auto;
         text-align: center;
     }
 }
