+++ src/main.jsx
@@ -1,11 +1,45 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import { BrowserRouter } from "react-router-dom";
+import { BrowserRouter, Routes, Route } from "react-router-dom";
+import { AuthProvider, useAuth } from "./context/AuthContext";
 import App from "./App";
+import FontLoader from "./Components/FontLoader/Fonts.jsx";
+import Navigatie from "./Components/Navigatie/Navigatie.jsx";
+import Home from "./Pages/Home.jsx";
+import About from "./Pages/About.jsx";
+import Contact from "./Pages/Contact.jsx";
+import Login from "./Pages/Login.jsx";
+import Register from "./Pages/Register.jsx";
+import NotFound from "./Pages/NotFound.jsx";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
-  <BrowserRouter>
-    <App />
-  </BrowserRouter>,
+    <React.StrictMode>
+        <AuthProvider>
+            <BrowserRouter>
+                <FontLoader />
+                <Navigatie />
+                <Routes>
+                    <Route path="/" element={<Home />} />
+                    <Route path="/about" element={<About />} />
+                    <Route path="/contact" element={<Contact />} />
+                    <Route path="/login" element={<Login />} />
+                    <Route
+                        path="/register"
+                        element={
+                            <PrivateRoute>
+                                <Register />
+                            </PrivateRoute>
+                        }
+                    />
+                    <Route path="*" element={<NotFound />} />
+                </Routes>
+            </BrowserRouter>
+        </AuthProvider>
+    </React.StrictMode>
 );
+
+function PrivateRoute({ children }) {
+    const { isLoggedIn } = useAuth();
+    return isLoggedIn ? children : <Login />;
+}
